data/token_price/requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
supabase>=1.0.0
influxdb-client>=1.36.0
asyncio>=3.4.3 